+++ src/components/checkbox/checkbox-base.ts
@@ -1,10 +1,10 @@
 import { LitElement } from 'lit';
 import { property, query, queryAssignedNodes, state } from 'lit/decorators.js';
 
+import { addKeyboardFocusRing } from '../common/controllers/focus-ring.js';
 import { alternateName } from '../common/decorators/alternateName.js';
 import { blazorDeepImport } from '../common/decorators/blazorDeepImport.js';
 import { blazorTwoWayBind } from '../common/decorators/blazorTwoWayBind.js';
-import { watch } from '../common/decorators/watch.js';
 import type { Constructor } from '../common/mixins/constructor.js';
 import { EventEmitterMixin } from '../common/mixins/event-emitter.js';
 import { FormAssociatedRequiredMixin } from '../common/mixins/form-associated-required.js';
@@ -25,17 +25,15 @@ export class IgcCheckboxBaseComponent extends FormAssociatedRequiredMixin(
 ) {
   protected override validators: Validator<this>[] = [requiredBooleanValidator];
 
+  protected _kbFocus = addKeyboardFocusRing(this);
   protected _value!: string;
   protected _checked = false;
 
-  @query('input[type="checkbox"]', true)
+  @query('input', true)
   protected input!: HTMLInputElement;
 
   @queryAssignedNodes({ flatten: true })
   protected label!: Array<Node>;
-
-  @state()
-  protected focused = false;
 
   @state()
   protected hideLabel = false;
@@ -66,10 +64,7 @@ export class IgcCheckboxBaseComponent extends FormAssociatedRequiredMixin(
     this._checked = Boolean(value);
     this.setFormValue(this._checked ? this.value || 'on' : null);
     this.updateValidity();
-
-    if (this.hasUpdated) {
-      this.setInvalidState();
-    }
+    this.setInvalidState();
   }
 
   public get checked(): boolean {
@@ -83,20 +78,17 @@ export class IgcCheckboxBaseComponent extends FormAssociatedRequiredMixin(
   @property({ reflect: true, attribute: 'label-position' })
   public labelPosition: 'before' | 'after' = 'after';
 
-  constructor() {
-    super();
-    this.addEventListener('keyup', this.handleKeyUp);
+  protected override createRenderRoot() {
+    const root = super.createRenderRoot();
+    root.addEventListener('slotchange', () => {
+      this.hideLabel = this.label.length < 1;
+    });
+    return root;
   }
 
   public override connectedCallback() {
     super.connectedCallback();
     this.updateValidity();
-  }
-
-  @watch('focused', { waitUntilFirstUpdate: true })
-  @watch('indeterminate', { waitUntilFirstUpdate: true })
-  protected handleChange() {
-    this.invalid = !this.checkValidity();
   }
 
   /** Simulates a click on the control. */
@@ -123,27 +115,11 @@ export class IgcCheckboxBaseComponent extends FormAssociatedRequiredMixin(
 
   protected handleBlur() {
     this.emitEvent('igcBlur');
-    this.focused = false;
+    this._kbFocus.reset();
   }
 
   protected handleFocus() {
     this._dirty = true;
     this.emitEvent('igcFocus');
-  }
-
-  protected handleMouseDown(event: PointerEvent) {
-    event.preventDefault();
-    this.input.focus();
-    this.focused = false;
-  }
-
-  protected handleKeyUp() {
-    if (!this.focused) {
-      this.focused = true;
-    }
-  }
-
-  protected handleSlotChange() {
-    this.hideLabel = this.label.length < 1;
   }
 }