+++ src/components/checkbox/switch.ts
@@ -48,7 +48,7 @@ export default class IgcSwitchComponent extends IgcCheckboxBaseComponent {
       <label
         part=${partNameMap({ base: true, checked: this.checked })}
         for=${this.inputId}
-        @pointerdown=${this.handleMouseDown}
+        @pointerdown=${this._kbFocus.reset}
       >
         <input
           id=${this.inputId}
@@ -69,7 +69,7 @@ export default class IgcSwitchComponent extends IgcCheckboxBaseComponent {
           part=${partNameMap({
             control: true,
             checked: this.checked,
-            focused: this.focused,
+            focused: this._kbFocus.focused,
           })}
         >
           <span
@@ -81,7 +81,7 @@ export default class IgcSwitchComponent extends IgcCheckboxBaseComponent {
           part=${partNameMap({ label: true, checked: this.checked })}
           id=${this.labelId}
         >
-          <slot @slotchange=${this.handleSlotChange}></slot>
+          <slot></slot>
         </span>
       </label>
     `;