+++ src/components/common/decorators/blazorIndirectRender.ts
@@ -1,4 +1,7 @@
-import type { AbstractConstructor, Constructor } from '../mixins/constructor';
+import type {
+  AbstractConstructor,
+  Constructor,
+} from '../mixins/constructor.js';
 
 /**
  * Indicates a class should use the indirect renderer in Blazor.