+++ src/components/common/util.ts
@@ -8,6 +8,8 @@ export const partNameMap = (partNameInfo: PartNameInfo) => {
     .join(' ');
 };
 
+export function noop() {}
+
 export const asPercent = (part: number, whole: number) => (part / whole) * 100;
 
 export const clamp = (number: number, min: number, max: number) =>
@@ -49,8 +51,11 @@ export function createCounter() {
   };
 }
 
+/**
+ * Returns whether an element has a Left-to-Right directionality.
+ */
 export function isLTR(element: HTMLElement) {
-  return getComputedStyle(element).getPropertyValue('direction') === 'ltr';
+  return element.matches(':dir(ltr)');
 }
 
 /**