+++ src/components/radio/radio.ts
@@ -4,6 +4,7 @@ import { ifDefined } from 'lit/directives/if-defined.js';
 import { live } from 'lit/directives/live.js';
 
 import { themes } from '../../theming/theming-decorator.js';
+import { addKeyboardFocusRing } from '../common/controllers/focus-ring.js';
 import {
   addKeybindings,
   arrowDown,
@@ -64,7 +65,8 @@ export default class IgcRadioComponent extends FormAssociatedRequiredMixin(
       key: 'valueMissing',
       message: messages.required,
       isValid: () => {
-        const { radios, checked } = this.group;
+        const radios = this._radios;
+        const checked = this._checkedRadios;
         return radios.some((radio) => radio.required)
           ? checked.length > 0
           : true;
@@ -74,6 +76,7 @@ export default class IgcRadioComponent extends FormAssociatedRequiredMixin(
 
   private inputId = `radio-${IgcRadioComponent.increment()}`;
   private labelId = `radio-label-${this.inputId}`;
+  private _kbFocus = addKeyboardFocusRing(this);
 
   protected _checked = false;
   protected _value!: string;
@@ -88,17 +91,30 @@ export default class IgcRadioComponent extends FormAssociatedRequiredMixin(
   private _tabIndex = 0;
 
   @state()
-  private focused = false;
-
-  @state()
   protected hideLabel = false;
 
-  private get group() {
-    return getGroup(this);
+  /** Returns all radio elements from the group, that is having the same name property. */
+  private get _radios() {
+    return getGroup(this).radios;
+  }
+
+  /** All sibling radio elements of the one invoking the getter. */
+  private get _siblings() {
+    return getGroup(this).siblings;
+  }
+
+  /** All non-disabled radio elements from the group. */
+  private get _active() {
+    return getGroup(this).active;
+  }
+
+  /** All checked radio elements from the group. */
+  private get _checkedRadios() {
+    return getGroup(this).checked;
   }
 
   protected override setDefaultValue(): void {
-    this._defaultValue = this === this.group.checked[0];
+    this._defaultValue = this === this._checkedRadios[0];
   }
 
   /**
@@ -125,10 +141,7 @@ export default class IgcRadioComponent extends FormAssociatedRequiredMixin(
   @blazorTwoWayBind('igcChange', 'detail')
   public set checked(value: boolean) {
     this._checked = Boolean(value);
-
-    if (this.hasUpdated) {
-      this._checked ? this._updateCheckedState() : this._updateUncheckedState();
-    }
+    this._checked ? this._updateCheckedState() : this._updateUncheckedState();
   }
 
   public get checked(): boolean {
@@ -144,7 +157,6 @@ export default class IgcRadioComponent extends FormAssociatedRequiredMixin(
 
   constructor() {
     super();
-    this.addEventListener('keyup', this.handleKeyUp);
 
     addKeybindings(this, {
       skip: () => this.disabled,
@@ -156,10 +168,18 @@ export default class IgcRadioComponent extends FormAssociatedRequiredMixin(
       .set(arrowDown, () => this.navigate(1));
   }
 
+  protected override createRenderRoot() {
+    const root = super.createRenderRoot();
+    root.addEventListener('slotchange', () => {
+      this.hideLabel = this.label.length < 1;
+    });
+    return root;
+  }
+
   public override connectedCallback() {
     super.connectedCallback();
 
-    this._checked = this === this.group.checked[0];
+    this._checked = this === this._checkedRadios[0];
     this.updateValidity();
   }
 
@@ -185,7 +205,7 @@ export default class IgcRadioComponent extends FormAssociatedRequiredMixin(
    * As long as `message` is not empty, the control is considered invalid.
    */
   public override setCustomValidity(message: string): void {
-    const { radios } = this.group;
+    const radios = this._radios;
 
     for (const radio of radios) {
       radio.updateValidity(message);
@@ -195,7 +215,7 @@ export default class IgcRadioComponent extends FormAssociatedRequiredMixin(
 
   @watch('required', { waitUntilFirstUpdate: true })
   protected override requiredChange(): void {
-    const { radios } = this.group;
+    const radios = this._radios;
 
     for (const radio of radios) {
       radio.updateValidity();
@@ -204,7 +224,7 @@ export default class IgcRadioComponent extends FormAssociatedRequiredMixin(
   }
 
   private _updateCheckedState() {
-    const { siblings } = this.group;
+    const siblings = this._siblings;
 
     this.setFormValue(this.value || 'on');
     this.updateValidity();
@@ -222,25 +242,30 @@ export default class IgcRadioComponent extends FormAssociatedRequiredMixin(
   }
 
   private _updateUncheckedState() {
-    const { siblings } = this.group;
+    const siblings = this._siblings;
 
     this.setFormValue(null);
     this.updateValidity();
     this.setInvalidState();
 
-    if (this.hasUpdated) {
-      this._tabIndex = -1;
-    }
+    this._tabIndex = -1;
 
     for (const radio of siblings) {
       radio.updateValidity();
     }
   }
 
-  protected handleMouseDown(event: PointerEvent) {
-    event.preventDefault();
-    this.input.focus();
-    this.focused = false;
+  protected override formResetCallback() {
+    super.formResetCallback();
+    this._resetTabIndexes();
+  }
+
+  /** Called after a form reset callback to restore default keyboard navigation. */
+  private _resetTabIndexes() {
+    const radios = this._radios;
+    for (const radio of radios) {
+      radio._tabIndex = 0;
+    }
   }
 
   protected handleClick() {
@@ -250,21 +275,15 @@ export default class IgcRadioComponent extends FormAssociatedRequiredMixin(
 
   protected handleBlur() {
     this.emitEvent('igcBlur');
-    this.focused = false;
+    this._kbFocus.reset();
   }
 
   protected handleFocus() {
     this.emitEvent('igcFocus');
   }
 
-  protected handleKeyUp() {
-    if (!this.focused) {
-      this.focused = true;
-    }
-  }
-
   protected navigate(idx: number) {
-    const { active } = this.group;
+    const active = this._active;
     const nextIdx = wrap(0, active.length - 1, active.indexOf(this) + idx);
     const radio = active[nextIdx];
 
@@ -273,22 +292,19 @@ export default class IgcRadioComponent extends FormAssociatedRequiredMixin(
     radio.emitEvent('igcChange', { detail: radio.checked });
   }
 
-  protected handleSlotChange() {
-    this.hideLabel = this.label.length < 1;
-  }
-
   protected override render() {
     const labelledBy = this.getAttribute('aria-labelledby');
+    const checked = this.checked;
 
     return html`
       <label
         part=${partNameMap({
           base: true,
-          checked: this.checked,
-          focused: this.focused,
+          checked,
+          focused: this._kbFocus.focused,
         })}
         for=${this.inputId}
-        @pointerdown=${this.handleMouseDown}
+        @pointerdown=${this._kbFocus.reset}
       >
         <input
           id=${this.inputId}
@@ -297,27 +313,27 @@ export default class IgcRadioComponent extends FormAssociatedRequiredMixin(
           value=${ifDefined(this.value)}
           .required=${this.required}
           .disabled=${this.disabled}
-          .checked=${live(this.checked)}
+          .checked=${live(checked)}
           tabindex=${this._tabIndex}
-          aria-checked=${this.checked ? 'true' : 'false'}
+          aria-checked=${checked ? 'true' : 'false'}
           aria-disabled=${this.disabled ? 'true' : 'false'}
           aria-labelledby=${labelledBy ? labelledBy : this.labelId}
           @click=${this.handleClick}
           @blur=${this.handleBlur}
           @focus=${this.handleFocus}
         />
-        <span part=${partNameMap({ control: true, checked: this.checked })}>
+        <span part=${partNameMap({ control: true, checked })}>
           <span
             .hidden=${this.disabled}
-            part=${partNameMap({ ripple: true, checked: this.checked })}
+            part=${partNameMap({ ripple: true, checked })}
           ></span>
         </span>
         <span
           .hidden=${this.hideLabel}
-          part=${partNameMap({ label: true, checked: this.checked })}
+          part=${partNameMap({ label: true, checked })}
           id=${this.labelId}
         >
-          <slot @slotchange=${this.handleSlotChange}></slot>
+          <slot></slot>
         </span>
       </label>
     `;